ml/vertex/predictor/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
google-cloud-storage
numpy
xgboost
scikit-learn
